Use the v0.34 definition of `abci.Event` which does not enforce valid UTF-8 data
Use the v0.34 definition of `abci.Event` which does not enforce valid UTF-8 data
for its `key` and `value` attributes, specifying them as `bytes` instead of `string`.
This is required, because ibc-go emits event attributes which are not valid UTF-8,
so we need to use this definition to be able to parse them.
In Protobuf, `bytes` and `string` are wire-compatible, so doing this strictly
increases the amount fo data we can parse.